pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
bs4 = "^0.0.1"
fastapi = "^0.104.1"
openai = "^1.2.3"
pydantic = "^2.4.2"
uvicorn = "^0.15.0"
langchain = "^0.0.334"
tiktoken = "^0.5.1"
xds-protos = "^1.59.2"
google-api-python-client = "^2.107.0"
google-auth-oauthlib = "^1.1.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"